data-science/requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3

# Data Analysis
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# Database
pymongo==4.6.0
python-dotenv==1.0.0

# Data Processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2023.3
